phase2_embedding_indexing/requirements.txt
pinecone-client>=3.0.0
sentence-transformers>=2.2.2
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
